requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.3.0
asgiref==3.8.1
async-timeout==4.0.3
attrs==23.2.0
boto3==1.34.88
botocore==1.34.88
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
coreapi==2.3.3
coreschema==0.0.4
cryptography==42.0.5
dataclasses-json==0.6.4
decorator==4.4.2
distro==1.9.0
Django==5.0.4
django-allow-cidr==0.7.1
django-rest-swagger==2.2.0
djangorestframework==3.15.1
exceptiongroup==1.2.1
filelock==3.13.4
frozenlist==1.4.1
fsspec==2024.3.1
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
imageio==2.34.0
imageio-ffmpeg==0.4.9
itypes==1.2.0
Jinja2==3.1.3
jmespath==1.0.1
jsonpatch==1.33
jsonpointer==2.4
langchain==0.1.16
langchain-community==0.0.34
langchain-core==0.1.45
langchain-text-splitters==0.0.1
langsmith==0.1.49
llvmlite==0.42.0
lxml==5.2.1
MarkupSafe==2.1.5
marshmallow==3.21.1
more-itertools==10.2.0
moviepy==1.0.3
mpmath==1.3.0
multidict==6.0.5
mypy-extensions==1.0.0
networkx==3.3
numba==0.59.1
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.19.3
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.1.105
openai==1.23.2
openai-whisper @ git+https://github.com/openai/whisper.git@ba3f3cd54b0e5b8ce1ab3de13e32122d0d5f98ab
openapi-codec==1.3.2
orjson==3.10.1
packaging==23.2
pdfminer.six==20231228
pdfplumber==0.11.0
pillow==10.3.0
proglog==0.1.10
psycopg2-binary==2.9.9
pycparser==2.22
pydantic==2.7.0
pydantic_core==2.18.1
pydub==0.25.1
pypdfium2==4.29.0
python-dateutil==2.9.0.post0
python-docx==1.1.0
python-dotenv==1.0.1
PyYAML==6.0.1
regex==2024.4.16
requests==2.31.0
s3transfer==0.10.1
simplejson==3.19.2
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.29
sqlparse==0.5.0
sympy==1.12
tenacity==8.2.3
tiktoken==0.6.0
torch==2.2.2
tqdm==4.66.2
triton==2.2.0
typing-inspect==0.9.0
typing_extensions==4.11.0
uritemplate==4.1.1
urllib3==2.2.1
whitenoise==6.6.0
yarl==1.9.4
